todo: frag crash

+++ analytical_engine/java/giraph-on-grape/src/main/java/com/alibaba/graphscope/loader/impl/FileLoader.java
@@ -202,7 +202,7 @@ public class FileLoader implements LoaderBase {
 
         long sum = 0;
         for (int i = 0; i < threadNum; ++i) {
-            sum += (int) futures[i].get();
+            sum += (Long) futures[i].get();
         }
         logger.info("worker {} loaded {} lines ", workerId, sum);
     }
